Add single job class
